This uses a mix of ceiling and floor instead of rounding.
Rounding was getting incorrect results so This is an attempt to fix.

I tried always ceiling and was also getting wrong results.

So using excel I looked at the math. Saw the average was less than amount needed to round up. So I am ceiling that. And flooring the first waifu in the party.